added profile tests
